Some multithreaded optimizations
Problem: running the MapReduce coreset is slower with more threads than
it is with a single one.

Observing the output of htop, it seems that a lot of time during the
call to kcenter is spent doing system calls, which is slower when using
multiple threads. I have still to pinpoint the offending part of the
code.

So far, switching from iterators to loops with explicit indexing made
the situation slightly better.